refactor: mvc 패턴 패키지 구조 적용
