+++ src/main/java/me/namuhuchutong/step4/RacingCarApplication.java
@@ -1,5 +1,11 @@
 package me.namuhuchutong.step4;
 
+import me.namuhuchutong.step4.controller.RacingCar;
+import me.namuhuchutong.step4.domain.RacingCarService;
+import me.namuhuchutong.step4.domain.rule.RandomRacingRule;
+import me.namuhuchutong.step4.view.InputView;
+import me.namuhuchutong.step4.view.OutputView;
+
 public class RacingCarApplication {
 
     public static void main(String[] args){
