+++ src/main/java/me/namuhuchutong/step4/domain/CarNames.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.step4;
+package me.namuhuchutong.step4.domain;
 
 import java.util.List;
