+++ src/main/java/me/namuhuchutong/step4/domain/RacingCarService.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.step4;
+package me.namuhuchutong.step4.domain;
 
 import static java.util.Comparator.*;
 import static java.util.stream.Collectors.*;
@@ -6,6 +6,9 @@ import static java.util.stream.Collectors.*;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Stream;
+import me.namuhuchutong.step4.domain.rule.RacingRule;
+import me.namuhuchutong.step4.dto.RacingResult;
+import me.namuhuchutong.step4.dto.UserInputInformation;
 
 public class RacingCarService {
