+++ src/main/java/me/namuhuchutong/step4/domain/rule/RacingRule.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.step4;
+package me.namuhuchutong.step4.domain.rule;
 
 @FunctionalInterface
 public interface RacingRule {
