+++ src/main/java/me/namuhuchutong/step4/dto/UserRequest.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.step4;
+package me.namuhuchutong.step4.dto;
 
 public class UserRequest {
