+++ src/test/java/me/namuhuchutong/step4/UserInputInterceptorTest.java
@@ -3,6 +3,9 @@ package me.namuhuchutong.step4;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.stream.Stream;
+import me.namuhuchutong.step4.controller.UserInputInterceptor;
+import me.namuhuchutong.step4.dto.UserInputInformation;
+import me.namuhuchutong.step4.dto.UserRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
